refactor: extract MenuTriggerButton into own file
